app/requirements.txt
fastapi
uvicorn
pydantic>=2
requests
playwright
python-dotenv